% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.phyloseq.R
\name{as.phyloseq}
\alias{as.phyloseq}
\title{Coerce a SummarizedExperiment class object to a phyloseq class object}
\usage{
as.phyloseq(x)
}
\arguments{
\item{x}{A SummarizedExperiment object from the HMP16SData package}
}
\value{
a phyloseq object
}
\description{
Coerce a SummarizedExperiment class object to a phyloseq class object
}
