% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_phyloseq.R
\name{as_phyloseq}
\alias{as_phyloseq}
\title{Coerce a SummarizedExperiment object to a phyloseq object}
\usage{
as_phyloseq(x)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} object
from the \code{\link{HMP16SData}} package}
}
\value{
A \code{\link[phyloseq]{phyloseq-class}} class object
}
\description{
Description
}
