% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{A data dictionary describing dbGaP metadata variables}
\format{A \code{\link[base]{data.frame}} with 789 rows and 8 variables:
\describe{
  \item{Variable}{the name of the variable as it appears in
  \code{\link[SummarizedExperiment]{SummarizedExperiment}} objects}
  \item{Description}{a description of the variable}
  \item{Type}{the type of variable, as specified by dbGaP}
  \item{Units}{the units of the variable (e.g. days)}
  \item{Minimum}{the smallest possible value of the variable}
  \item{Maximum}{the largest possible value of the variable}
  \item{Coded values}{the encoding of coded categorical variables, if needed}
  \item{Comment}{any comments from dbGaP, most pertain to gated fields}
}}
\source{
\url{https://tinyurl.com/y7rgelao}

\url{https://tinyurl.com/y7at85tv}

\url{https://tinyurl.com/y7af3arf}

\url{https://tinyurl.com/ych4wywl}

\url{https://tinyurl.com/y9hwfn9e}

\url{https://tinyurl.com/y7jtghxv}

\url{https://tinyurl.com/y7znjmak}
}
\usage{
dictionary
}
\description{
A data dictionary containing descriptions of the variables attached using the
\code{\link{attach_dbGaP}} function, as translated from dbGaP XLM data
dictionary files.
}
\examples{
data("dictionary")
}
\seealso{
\code{\link{attach_dbGaP}}
}
\keyword{datasets}
