% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMP16SData.R
\name{HMP16SData}
\alias{HMP16SData}
\title{16S rRNA sequencing data from the Human Microbiome Project}
\description{
HMP16SData is a Bioconductor ExperimentData package of the Human Microbiome
Project (HMP) 16S rRNA sequencing data for variable regions 1–3 and 3–5. Raw
data files are provided in the package as downloaded from the HMP Data
Analysis and Coordination Center website. Processed data is provided as
\code{\link[SummarizedExperiment]{SummarizedExperiment}} objects via
\code{\link[ExperimentHub]{ExperimentHub}}.
}
