% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_one.R
\name{kable_one}
\alias{kable_one}
\title{Produce a summary HTML table of key demographic variables}
\usage{
kable_one(x)
}
\arguments{
\item{x}{A \code{data.frame} object or a named \code{list} of
\code{data.frame} objects from the output of \code{\link{table_one}}}
}
\value{
A summary HTML table of key demographic variables
}
\description{
Sometimes it is desirable to produce a summary of key demographic variables..
}
\examples{
V13() \%>\%
    table_one() \%>\%
    kable_one()

}
\seealso{
\code{\link{table_one}}
}
