% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one.R
\name{table_one}
\alias{table_one}
\title{Title}
\usage{
table_one(foo, visit_number = FALSE, sex = TRUE, run_center = FALSE,
  hmp_body_subsite = TRUE)
}
\arguments{
\item{foo}{description}

\item{visit_number}{description}

\item{sex}{description}

\item{run_center}{description}

\item{hmp_body_subsite}{description}
}
\description{
Title
}
