% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V35.R
\name{V35}
\alias{V35}
\title{16S rRNA sequencing data for variable regions 3-5}
\format{A SummarizedExperiment with 45,383 features and 4,743 samples:
\describe{
    \item{random_subject_identifier}{a random subject identifier}
    \item{visit_number}{visit number, between 1 and 3}
    \item{sex}{sex, female or male}
    \item{run_center}{center where sample sequencing took place: Baylor
    College of Medicine (BCM), the Broad Institute (BI), the J. Craig Venter
    Institute (JCVI), or the Genome Sequencing Center at Washington
    University (WUGC)}
    \item{hmp_body_subsite}{body subsite where sample was collected}
    \item{description}{long description providing a random sample ID number,
    a random subject ID number, and sex}
    \item{consensus_lineage}{the most detailed lineage description shared by
    the sequences within an OTU}
}}
\source{
\url{http://downloads.ihmpdcc.org/data/HMQCP/otu_table_psn_v35.txt.gz}

\url{http://downloads.ihmpdcc.org/data/HMQCP/v35_map_uniquebyPSN.txt.bz2}
}
\arguments{
\item{metadata}{logical; if TRUE only the metadata is downloaded, rather than
the entire resource}
}
\description{
16S rRNA sequencing was performed on samples collected at five major body
sites from 300 healthy adults between the ages of 18 and 40
}
\examples{
V35()

}
