% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_phyloseq.R
\name{as_phyloseq}
\alias{as_phyloseq}
\title{Coerce a SummarizedExperiment class object to a phyloseq class object}
\usage{
as_phyloseq(x)
}
\arguments{
\item{x}{a SummarizedExperiment object from the HMP16SData package}
}
\value{
a phyloseq object
}
\description{
Coerce a SummarizedExperiment class object to a phyloseq class object
}
\examples{
V13() \%>\% as.phyloseq()

}
