% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_dbGaP.R
\name{attach_dbGaP}
\alias{attach_dbGaP}
\title{Attach dbGaP metadata to HMP16SData SummarizedExperiment objects}
\usage{
attach_dbGaP(x, dbGaP_repository_key)
}
\arguments{
\item{x}{a SummarizedExperiment object from the HMP16SData package}

\item{dbGaP_repository_key}{a repository key downloaded from dbGaP}
}
\value{
a SummarizedExperiment object with protected metadata from dbGaP
}
\description{
Attach dbGaP metadata to HMP16SData SummarizedExperiment objects
}
\examples{
\dontrun{V13() \%>\% attach_dbGaP("~/prj_12146.ngc")}

}
\seealso{
\code{\link{dictionary}}
}
