% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_dbGaP.R
\name{attach_dbGaP}
\alias{attach_dbGaP}
\title{Attach dbGaP metadata to HMP16SData SummarizedExperiment objects}
\usage{
attach_dbGaP(x, dbGaP_repository_key = "")
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} object
from the \code{\link{HMP16SData}} package}

\item{dbGaP_repository_key}{A repository key downloaded from dbGaP}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} object
with protected metadata from dbGaP
}
\description{
Description
}
\note{
Note
}
\examples{
\dontrun{
V13() \%>\%
    attach_dbGaP("~/prj_12146.ngc")
}

}
\seealso{
\code{\link{dictionary}}
}
