% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{A data dictionary describing dbGaP metadata variables}
\format{A data.frame with 787 rows and 8 variables:
\describe{
  \item{Variable}{the name of the variable as it appears in the
  SummarizedExperiment objects}
  \item{Description}{a description of the variable}
  \item{Type}{the type of variable, as specified by dbGaP}
  \item{Units}{the units of the variable (e.g. days)}
  \item{Minimum}{the smallest possible value of the variable}
  \item{Maximum}{the largest possible value of the variable}
  \item{Coded values}{the encoding of coded categorical variables, if needed}
  \item{Comment}{any comments from dbGaP, most pertain to gated fields}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/dbgap/studies/phs000228/phs000228.v4.p1/pheno_variable_summaries/phs000228.v4.pht001185.v4.EMMES_HMP_Sample.data_dict.xml}
\url{ftp://ftp.ncbi.nlm.nih.gov/dbgap/studies/phs000228/phs000228.v4.p1/pheno_variable_summaries/phs000228.v4.pht001193.v3.EMMES_HMP_GTV.data_dict.xml}
\url{ftp://ftp.ncbi.nlm.nih.gov/dbgap/studies/phs000228/phs000228.v4.p1/pheno_variable_summaries/phs000228.v4.pht002157.v1.EMMES_HMP_DTP_DHX_DVD.data_dict.xml}
\url{ftp://ftp.ncbi.nlm.nih.gov/dbgap/studies/phs000228/phs000228.v4.p1/pheno_variable_summaries/phs000228.v4.pht001187.v3.EMMES_HMP_DCM.data_dict.xml}
\url{ftp://ftp.ncbi.nlm.nih.gov/dbgap/studies/phs000228/phs000228.v4.p1/pheno_variable_summaries/phs000228.v4.pht001184.v3.EMMES_HMP_Subject.data_dict.xml}
\url{ftp://ftp.ncbi.nlm.nih.gov/dbgap/studies/phs000228/phs000228.v4.p1/pheno_variable_summaries/phs000228.v4.pht002158.v1.EMMES_HMP_DEM_ENR.data_dict.xml}
\url{ftp://ftp.ncbi.nlm.nih.gov/dbgap/studies/phs000228/phs000228.v4.p1/pheno_variable_summaries/phs000228.v4.pht002156.v1.EMMES_HMP_DSU.data_dict.xml}
}
\usage{
dictionary
}
\description{
A dictionary containing descriptions of the variables attached using the
\code{\link{attach_dbGaP}} function, as translated from NCBI XLM data
dictionary files.
}
\examples{
data("dictionary")

}
\seealso{
\code{\link{attach_dbGaP}}
}
\keyword{datasets}
