% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMP16SData-package.r
\docType{package}
\name{HMP16SData}
\alias{HMP16SData}
\alias{HMP16SData-package}
\title{16S rRNA Microbiome Data from the Human Microbiome Project}
\description{
The HMP16SData package is a Bioconductor ExperimentHub package of the Human
Microbiome Project (HMP) 16S rRNA sequencing data for variable regions 1-3
and 3-5. The raw data files are provided in the package as downloaded from
the HMP Data Analysis and Coordination Center website. Processed data is
provided as SummarizedExperiment objects via ExperimentHub.
}
