% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_one.R
\name{kable_one}
\alias{kable_one}
\title{Title}
\usage{
kable_one(x)
}
\arguments{
\item{x}{Description}
}
\value{
Description
}
\description{
Description
}
\examples{
V13() \%>\%
    table_one() \%>\%
    kable_one()

}
\seealso{
\code{\link{table_one}}
}
