% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one.R
\name{table_one}
\alias{table_one}
\title{Title}
\usage{
table_one(x, VISITNO = TRUE, SEX = TRUE, RUN_CENTER = TRUE,
  HMP_BODY_SITE = TRUE, HMP_BODY_SUBSITE = TRUE)
}
\arguments{
\item{x}{Description}

\item{VISITNO}{Description}

\item{SEX}{Description}

\item{RUN_CENTER}{Description}

\item{HMP_BODY_SITE}{Description}

\item{HMP_BODY_SUBSITE}{Description}
}
\value{
Description
}
\description{
Description
}
\examples{
V13() \%>\%
    table_one()

}
\seealso{
\code{\link{kable_one}}
}
