% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one.R
\name{table_one}
\alias{table_one}
\title{Produce a summary data.frame of key demographic variables}
\usage{
table_one(x, VISITNO = TRUE, SEX = TRUE, RUN_CENTER = TRUE,
  HMP_BODY_SITE = TRUE, HMP_BODY_SUBSITE = TRUE)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
object or a named \code{list} of
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}} objects from
the \code{\link{HMP16SData}} package}

\item{VISITNO}{logical; if \code{FALSE}, the \code{VISITNO} column(s) of
\code{SummarizedExperiment} \code{colData} will not be summarized}

\item{SEX}{logical; if \code{FALSE}, the \code{SEX} column(s) of
\code{SummarizedExperiment} \code{colData} will not be summarized}

\item{RUN_CENTER}{logical; if \code{FALSE}, the \code{RUN_CENTER} column(s)
of \code{SummarizedExperiment} \code{colData} will not be summarized}

\item{HMP_BODY_SITE}{logical; if \code{FALSE}, the \code{HMP_BODY_SITE}
column(s) of \code{SummarizedExperiment} \code{colData} will not be
summarized}

\item{HMP_BODY_SUBSITE}{logical; if \code{FALSE}, the \code{HMP_BODY_SUBSITE}
column(s) of \code{SummarizedExperiment} \code{colData} will not be
summarized}
}
\value{
A \code{data.frame} object or a named \code{list} of
\code{data.frame} objects
}
\description{
Sometimes it is desirable to produce a summary of key demographic variables
for presentation. The \code{\link{table_one}} and \code{\link{kable_one}}
methods are a quick way to do so – they not only summarize key demographic
variables from \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
object(s) in the \code{\link{HMP16SData}} package but remove abbreviations
and underscores in column and variable names that might otherwise be
ambigious. The \code{\link{table_one}} method returns a \emph{tidy} (i.e. one
sample observation per line) \code{data.frame} object or a named \code{list}
of \emph{tidy} \code{data.frame} objects. The \code{\link{kable_one}} method
can then be used to produce a publication-ready \code{HTML} table that could,
for example, be pasted into a word processor.
}
\examples{
V13() \%>\%
    table_one()

}
\seealso{
\code{\link{kable_one}}
}
