% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_identifiers.R
\name{translate_identifiers}
\alias{translate_identifiers}
\title{Translate taxonomic identifiers to NCBI taxonomy standard}
\usage{
translate_identifiers(x)
}
\arguments{
\item{x}{a taxonomic identifier to translate to NCBI taxonomy standard}
}
\value{
a taxonomic identifier translated to NCBI taxonomy standard
}
\description{
Translate taxonomic identifiers to NCBI taxonomy standard
}
\examples{
translate_identifiers("f__LaCtObaCiLlAcEaE;g__LaCTobaCIllus")

}
