% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V35.R
\name{V35}
\alias{V35}
\title{16S rRNA sequencing data for variable regions 3–5}
\format{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} with
45,383 features and 4,743 samples:
\subsection{colData}{
    \describe{
        \item{RSID}{a random subject identifier}
        \item{VISITNO}{visit number, between 1 and 3}
        \item{SEX}{sex, female or male}
        \item{RUN_CENTER}{center where sample sequencing took place: Baylor
        College of Medicine (BCM), the Broad Institute (BI), the J. Craig
        Venter Institute (JCVI), or the Genome Sequencing Center at
        Washington University (WUGC)}
        \item{HMP_BODY_SUBSITE}{body subsite where the sample was collected}
    }
}
\subsection{rowData}{
    \describe{
        \item{CONSENSUS_LINEAGE}{the most detailed lineage description shared
        by the sequences within an OTU}
        \item{SUPERKINGDOM}{superkingdom taxonomy, assumed to be Bacteria}
        \item{PHYLUM}{phylum taxonomy parsed from \code{CONSENSUS_LINEAGE}}
        \item{CLASS}{calss taxonomy parsed from \code{CONSENSUS_LINEAGE}}
        \item{ORDER}{order taxonomy parsed from \code{CONSENSUS_LINEAGE}}
        \item{FAMILY}{family taxonomy parsed from \code{CONSENSUS_LINEAGE}}
        \item{GENUS}{genus taxonomy parsed from \code{CONSENSUS_LINEAGE}}
    }
}}
\source{
\url{https://tinyurl.com/y9rbpjl7}

\url{https://tinyurl.com/yaz35f22}
}
\usage{
V35(metadata = FALSE)
}
\arguments{
\item{metadata}{logical; if \code{TRUE} only the metadata is downloaded,
rather than the entire resource}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} object
}
\description{
16S rRNA sequencing was performed on samples collected at five major body
sites from 300 healthy adults between the ages of 18 and 40
}
\examples{
V35()

}
